chore: rename folder to keep name pattern
