+++ src/app/modules/animals/main/factories/use-cases/animal-childbirths-use-cases/remote-delete-animal-childbirth-use-case-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiHttpClient } from '@/core/main/factories/http'
 
-import { RemoteDeleteAnimalChildbirthUseCase } from '../../../../data/use-cases/childbirths-use-cases'
+import { RemoteDeleteAnimalChildbirthUseCase } from '../../../../data/use-cases/animal-childbirths-use-cases'
 
 import type { DeleteAnimalChildbirthUseCase } from '../../../../domain/use-cases/animal-childbirths-use-cases'
