+++ src/app/modules/animals/main/factories/use-cases/animal-childbirths-use-cases/remote-get-animal-childbirth-use-case-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiHttpClient } from '@/core/main/factories/http'
 
-import { RemoteGetAnimalChildbirthUseCase } from '../../../../data/use-cases/childbirths-use-cases'
+import { RemoteGetAnimalChildbirthUseCase } from '../../../../data/use-cases/animal-childbirths-use-cases'
 
 import type { AnimalChildbirthDetailsModel } from '../../../../domain/models/animal-childbirths-model'
 import type { GetAnimalChildbirthUseCase } from '../../../../domain/use-cases/animal-childbirths-use-cases'
