+++ src/app/modules/animals/main/factories/use-cases/animal-childbirths-use-cases/remote-get-animal-childbirths-use-case-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiHttpClient } from '@/core/main/factories/http'
 
-import { RemoteGetAnimalChildbirthsUseCase } from '../../../../data/use-cases/childbirths-use-cases'
+import { RemoteGetAnimalChildbirthsUseCase } from '../../../../data/use-cases/animal-childbirths-use-cases'
 
 import type { AnimalChildbirthModel } from '../../../../domain/models/animal-childbirths-model'
 import type { GetAnimalChildbirthsUseCase } from '../../../../domain/use-cases/animal-childbirths-use-cases'
