+++ src/app/modules/animals/main/factories/use-cases/animal-childbirths-use-cases/remote-update-animal-childbirth-use-case-factory.ts
@@ -1,6 +1,6 @@
 import { makeApiHttpClient } from '@/core/main/factories/http'
 
-import { RemoteUpdateAnimalChildbirthUseCase } from '../../../../data/use-cases/childbirths-use-cases'
+import { RemoteUpdateAnimalChildbirthUseCase } from '../../../../data/use-cases/animal-childbirths-use-cases'
 
 import type { UpdateAnimalChildbirthUseCase } from '../../../../domain/use-cases/animal-childbirths-use-cases'
